Handle the case that adds new l10n ports

+++ sync-out.sh
@@ -5,6 +5,10 @@ LO_WC=~/freebsd-ports-libreoffice
 
 for p in `cat list.txt`
 do
-        rm -fr ${SVN_WC}/${p}/*
+	if [ -d ${SVN_WC}/${p}/ ]; then
+		rm -fr ${SVN_WC}/${p}/*
+	else
+		mkdir -p ${SVN_WC}/${p}
+	fi
         cp -R ${LO_WC}/${p}/* ${SVN_WC}/${p}/
 done
